Move `CultureFeedAddressFactoryInterface` to `CultureFeed` folder
