+++ src/Address/CultureFeed/CultureFeedAddressFactoryInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Address;
+namespace CultuurNet\UDB3\Address\CultureFeed;
+
+use CultuurNet\UDB3\Address\Address;
 
 interface CultureFeedAddressFactoryInterface
 {
