+++ src/Address/CultureFeedAddressFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Address;
 
+use CultuurNet\UDB3\Address\CultureFeed\CultureFeedAddressFactoryInterface;
 use CultuurNet\UDB3\Model\ValueObject\Geography\CountryCode;
 
 class CultureFeedAddressFactory implements CultureFeedAddressFactoryInterface
